backend/pyproject.toml
[project]
name = "pwc-rag-backend"
version = "0.1.0"
description = "Financial Document RAG System"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pdfplumber>=0.11.0",
    "PyMuPDF>=1.24.0",
    "openai>=1.60.0",
    "pinecone>=5.0.0",
    "supabase>=2.0.0",
    "pydantic-settings>=2.0.0",
    "sse-starlette>=2.0.0",
    "tiktoken>=0.8.0",
    "python-multipart>=0.0.18",
    "azure-ai-documentintelligence>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
